refactor: move buildStaticFiles into cli-build, clean up exports
