+++ .changeset/swift-ravens-peel.md
@@ -1,0 +1,6 @@
+---
+"@sanity/cli-build": major
+"@sanity/cli": patch
+---
+
+Move buildStaticFiles into cli-build package and clean up exports.
