+++ packages/@sanity/cli-build/src/_exports/_internal/build.ts
@@ -1,5 +1,5 @@
-export {type AutoUpdatesBuildConfig} from '../../actions/build/autoUpdates.js'
 export {buildDebug} from '../../actions/build/buildDebug.js'
+export {buildStaticFiles} from '../../actions/build/buildStaticFiles.js'
 export {checkRequiredDependencies} from '../../actions/build/checkRequiredDependencies.js'
 export {checkStudioDependencyVersions} from '../../actions/build/checkStudioDependencyVersions.js'
 export {
@@ -7,17 +7,9 @@ export {
   getAutoUpdatesImportMap,
   getModuleUrl,
 } from '../../actions/build/getAutoUpdatesImportMap.js'
-export {
-  extendViteConfigWithUserConfig,
-  finalizeViteConfig,
-  getViteConfig,
-} from '../../actions/build/getViteConfig.js'
-export {
-  resolveVendorBuildConfig,
-  type VendorBuildConfig,
-} from '../../actions/build/resolveVendorBuildConfig.js'
-export {writeFavicons} from '../../actions/build/writeFavicons.js'
-export {resolveEntries, writeSanityRuntime} from '../../actions/build/writeSanityRuntime.js'
+export {extendViteConfigWithUserConfig, getViteConfig} from '../../actions/build/getViteConfig.js'
+export {resolveVendorBuildConfig} from '../../actions/build/resolveVendorBuildConfig.js'
+export {writeSanityRuntime} from '../../actions/build/writeSanityRuntime.js'
 export {SANITY_CACHE_DIR} from '../../constants.js'
 export {AppBuildTrace, StudioBuildTrace} from '../../telemetry/build.telemetry.js'
-export {copyDir} from '../../util/copyDir.js'
+export {formatModuleSizes, sortModulesBySize} from '../../util/moduleFormatUtils.js'
